requirements.txt
fastapi[standard]~=0.116.1
passlib[bcrypt]
sqlmodel~=0.0.24
psycopg[binary]
python-dotenv~=1.1.1